Clean up stale new sample data flag and remove old sample data controller
