+++ packages/front-end/components/GetStarted/ViewSampleDataButton.tsx
@@ -1,13 +1,11 @@
 import { ProjectInterface } from "shared/types/project";
 import { useRouter } from "next/router";
-import { useGrowthBook } from "@growthbook/growthbook-react";
 import { useDemoDataSourceProject } from "@/hooks/useDemoDataSourceProject";
 import track from "@/services/track";
 import { useDefinitions } from "@/services/DefinitionsContext";
 import { useAuth } from "@/services/auth";
 import usePermissionsUtil from "@/hooks/usePermissionsUtils";
 import Button from "@/ui/Button";
-import { isCloud } from "@/services/env";
 
 const ViewSampleDataButton = ({
   resource = "experiment",
@@ -23,8 +21,6 @@ const ViewSampleDataButton = ({
 
   const { mutateDefinitions } = useDefinitions();
 
-  const gb = useGrowthBook();
-
   const openSample = async () => {
     if (exists && demoExperimentId) {
       if (resource === "experiment") {
@@ -36,14 +32,9 @@ const ViewSampleDataButton = ({
       const res = await apiCall<{
         project: ProjectInterface;
         experimentId: string;
-      }>(
-        isCloud() && gb.isOn("new-sample-data")
-          ? "/demo-datasource-project/new"
-          : "/demo-datasource-project",
-        {
-          method: "POST",
-        },
-      );
+      }>("/demo-datasource-project", {
+        method: "POST",
+      });
       track("Create Sample Project", {
         source: "get-started",
       });
