+++ packages/front-end/components/HomePage/ExperimentsGetStarted.tsx
@@ -3,13 +3,8 @@ import { useRouter } from "next/router";
 import { FaArrowLeft } from "react-icons/fa";
 import { ProjectInterface } from "shared/types/project";
 import { getDemoDatasourceProjectIdForOrganization } from "shared/demo-datasource";
-import { useGrowthBook } from "@growthbook/growthbook-react";
 import { useDefinitions } from "@/services/DefinitionsContext";
-import {
-  envAllowsCreatingMetrics,
-  hasFileConfig,
-  isCloud,
-} from "@/services/env";
+import { envAllowsCreatingMetrics, hasFileConfig } from "@/services/env";
 import NewDataSourceForm from "@/components/Settings/NewDataSourceForm";
 import MetricForm from "@/components/Metrics/MetricForm";
 import { DocLink } from "@/components/DocLink";
@@ -57,8 +52,6 @@ const ExperimentsGetStarted = (): React.ReactElement => {
 
   const { apiCall } = useAuth();
 
-  const gb = useGrowthBook();
-
   const openSampleExperiment = async () => {
     if (demoDataSourceProjectId && demoExperimentId) {
       router.push(`/experiment/${demoExperimentId}`);
@@ -66,14 +59,9 @@ const ExperimentsGetStarted = (): React.ReactElement => {
       const res = await apiCall<{
         project: ProjectInterface;
         experimentId: string;
-      }>(
-        isCloud() && gb.isOn("new-sample-data")
-          ? "/demo-datasource-project/new"
-          : "/demo-datasource-project",
-        {
-          method: "POST",
-        },
-      );
+      }>("/demo-datasource-project", {
+        method: "POST",
+      });
       track("Create Sample Project", {
         source: "experiments-get-started",
       });
